Update plotting method
Old method plotted all y-values to the mean x-value.
This works if the x-trajectories are overall similar;
however this assumption becomes untrue once we consider perturbation.
The updated script interpolates the run-specific (x,y) trajectories
to get an accurate estimate of the y-value at every x-point.
This enables accurate plotting of the mean and std.